+++ integration-tests/src/test/java/org/web3j/protocol/scenarios/FastRawTransactionManagerIT.java
@@ -38,6 +38,7 @@ import org.web3j.tx.response.PollingTransactionReceiptProcessor;
 import org.web3j.tx.response.QueuingTransactionReceiptProcessor;
 import org.web3j.utils.Convert;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.web3j.tx.TransactionManager.DEFAULT_POLLING_ATTEMPTS_PER_TX_HASH;
@@ -94,6 +95,24 @@ public class FastRawTransactionManagerIT extends Scenario {
         }
 
         assertTrue(transactionReceipts.isEmpty());
+    }
+
+    @Test
+    public void testTransactionResetNonce() throws Exception {
+        FastRawTransactionManager transactionManager =
+                new FastRawTransactionManager(
+                        web3j,
+                        ALICE,
+                        new PollingTransactionReceiptProcessor(
+                                web3j, POLLING_FREQUENCY, DEFAULT_POLLING_ATTEMPTS_PER_TX_HASH));
+
+        Transfer transfer = new Transfer(web3j, transactionManager);
+        BigInteger gasPrice = transfer.requestCurrentGasPrice();
+
+        createTransaction(transfer, gasPrice).send();
+        createTransaction(transfer, gasPrice).send();
+        transactionManager.resetNonce();
+        createTransaction(transfer, gasPrice).send();
     }
 
     @Test